+++ src/extension.ts
@@ -5,12 +5,13 @@ export function activate(context: vscode.ExtensionContext) {
 
     let disposable = vscode.commands.registerCommand('insert-cursor-at-beginning-of-each-line-selected.insertCursors', () => {
         let editor = vscode.window.activeTextEditor;
+        if (!editor) {
+            return;
+        }
 
         let currentPosition = editor.selection.start.character;
         let firstLine = editor.selection.start.line;
         let lastLine = editor.selection.end.line;
-
-        let newPosition = new vscode.Position(firstLine, currentPosition);
 
         let selections = [];
         for (let lineNumber = firstLine; lineNumber <= lastLine; lineNumber++) {